requirements.txt
numpy
Pillow
guizero
